requirements.txt
alembic==0.9.6
Babel==2.5.1
bcrypt==3.1.4
blinker==1.4
cffi==1.11.2
click==6.7
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-User==0.6.19
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
passlib==1.7.1
psycopg2==2.7.3.1
pycparser==2.18
pycryptodome==3.4.7
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
six==1.11.0
speaklater==1.3
SQLAlchemy==1.1.14
Werkzeug==0.12.2
WTForms==2.1
alembic==0.9.6
Babel==2.5.1
bcrypt==3.1.4
blinker==1.4
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
flake8==3.5.0
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-User==0.6.19
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
passlib==1.7.1
pew==1.1.0
pipenv==8.3.1
psutil==5.3.1
psycopg2==2.7.3.1
pycodestyle==2.3.1
pycparser==2.18
pycryptodome==3.4.7
pyflakes==1.6.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
requests==2.18.4
six==1.11.0
speaklater==1.3
SQLAlchemy==1.1.14
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
Werkzeug==0.12.2
WTForms==2.1
alembic==0.9.6
Babel==2.5.1
bcrypt==3.1.4
blinker==1.4
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
flake8==3.5.0
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-User==0.6.19
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
passlib==1.7.1
petname==2.2
pew==1.1.0
pipenv==8.3.1
psutil==5.3.1
psycopg2==2.7.3.1
pycodestyle==2.3.1
pycparser==2.18
pycryptodome==3.4.7
pyflakes==1.6.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
requests==2.18.4
six==1.11.0
smsgateway.me-api==0.1
speaklater==1.3
SQLAlchemy==1.1.14
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
Werkzeug==0.12.2
WTForms==2.1
alembic==0.9.6
Babel==2.5.1
bcrypt==3.1.4
blinker==1.4
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
flake8==3.5.0
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-User==0.6.19
Flask-WTF==0.14.2
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
locustio==0.8.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
msgpack-python==0.4.8
passlib==1.7.1
petname==2.2
pew==1.1.0
pipenv==8.3.1
psutil==5.3.1
psycopg2==2.7.3.1
pycodestyle==2.3.1
pycparser==2.18
pycryptodome==3.4.7
pyflakes==1.6.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
pyzmq==16.0.3
requests==2.18.4
six==1.11.0
speaklater==1.3
SQLAlchemy==1.1.14
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
Werkzeug==0.12.2
WTForms==2.1
